添加大厅机制的处理②，欲待与客户端对接进行测试（部分已成功测试）

+++ src/main/java/com/gameserver/gd/service/DeckService.java
@@ -96,6 +96,8 @@ public class DeckService {
     public boolean DeleteDeck(String username){
         DeckExample deckExample = new DeckExample();
         deckExample.or().andUsernameEqualTo(username);
+        if (deckMapper.selectByExample(deckExample).size()==0)
+            return true;
         return deckMapper.deleteByExample(deckExample)>0;
     }
 }
